+++ IndividualCapstone/Models/ServiceType.cs
@@ -12,7 +12,7 @@ namespace IndividualCapstone.Models
         [Key]
         public int Id { get; set; }
         public string NameOfService { get; set; }
-        public double CostOfService { get; set; }
+        public decimal CostOfService { get; set; }
       
     }
 }
